+++ src/main/java/lapr/project/ui/PositionalMessageUI.java
@@ -13,11 +13,11 @@ public class PositionalMessageUI implements Runnable {
     public void run(){
 
 
-        long mmsi;
+        int mmsi;
         Scanner read = new Scanner(System.in);
 
         System.out.println("Which ship will you read the positional messages?");
-        mmsi = read.nextLong();
+        mmsi = read.nextInt();
 
         if(!positionalMessageController.getPositionalMessages(mmsi)){
             System.out.println(mmsi+" doesn't exist!");
